Benchmarking of noise filtering and normalization impact on network inference

This function takes four types of matrices as input which were generated with the same network inference tool (e.g. GENIE3), but had different input data to the network inference tool in the context of noise-filtering and quantile-normalization. It is assumed that the four matrices inserted into this function come from network inference runs with the following input data specifications:

Raw mapped RNAseq count data without noise-filtering or quantile-normalization applied (adj_mat_raw_no_noise_removed_and_no_qnorm)

Raw mapped RNAseq count data with noise-filtering applied but no quantile-normalization (adj_mat_raw_noise_removed_but_no_qnorm)

Raw mapped RNAseq count data without noise-filtering but with quantile-normalization applied (adj_mat_no_noise_removed_but_qnorm)

Raw mapped RNAseq count data with noise-filtering applied and with quantile-normalization applied (adj_mat_noise_removed_and_qnorm)

All relevant pairwise comparisons are then performed internally with network_dist_pairwise_genes based on gene-wise Hamming Distance or Jaccard Similarity Coefficient computations.

network_benchmark_noise_filtering(
  adj_mat_raw_no_noise_removed_and_no_qnorm,
  adj_mat_raw_noise_removed_but_no_qnorm,
  adj_mat_no_noise_removed_but_qnorm,
  adj_mat_noise_removed_and_qnorm,
  grn_tool,
  threshold = "median",
  dist_type = "hamming",
  print_message = TRUE
)

Arguments

adj_mat_raw_no_noise_removed_and_no_qnorm
a weighted adjacency matrix derived from a network inference program where no noise-filtering or quantile-normalization was applied to the input data.

adj_mat_raw_noise_removed_but_no_qnorm
a weighted adjacency matrix derived from a network inference program where noise-filtering but no quantile-normalization was applied to the input data.

adj_mat_no_noise_removed_but_qnorm
a weighted adjacency matrix derived from a network inference program where no noise-filtering but quantile-normalization was applied to the input data.

adj_mat_noise_removed_and_qnorm
a weighted adjacency matrix derived from a network inference program where noise-filtering and quantile-normalization were applied to the input data.

grn_tool
a character string specifying the gene regulatory network inference tool that was used to generate input matrices.

threshold
we recommended to use network_rescale before using this function. Re-scaling will transform all values into a range [0,100]. The threshold can either be a numeric balue in the interval [0,100] or a character string specifying the following methods for automatically determining the threshold based on the input data:

threshold = "median": compute the median over the entire input adj_mat and use this median value as threshold for defining all edge weights of a genes equal or below the median value as 0 and all values above the median value as 1.

See network_make_binary for details.

dist_type
a distance method that shall be applied on the binary values for each gene. Available options are:

dist_type = "hamming": computes the hamming.distance for each gene between the two input matrices

dist_type = "jaccard": computes the jaccard for each gene between the two input matrices

See network_dist_pairwise_genes for details.

print_message
shall messages be printed?

Author

Hajk-Georg Drost

Examples

# Benchmark GENIE3 inferred networks with raw, no_noise, and quantile_norm combinations
genie3_49_raw <- as.matrix(read.csv(
system.file("data/network_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_raw <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_qnorm_no_noise_removed <- as.matrix(read.csv(
system.file("data/network_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_qnorm <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

# Run Benchmark using Hamming distance
benchmark_hamming <- network_benchmark_noise_filtering(
genie3_49_raw,
genie3_49_noNoiseCM_raw,
genie3_49_qnorm_no_noise_removed,
genie3_49_noNoiseCM_qnorm,
dist_type = "hamming",
grn_tool = "GENIE3")

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [11.27] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [9.95] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [10.75] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [11.04] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> 

#> Comparison 1: 'adj_mat_raw_no_noise_removed_and_no_qnorm' vs 'adj_mat_raw_noise_removed_but_no_qnorm'

#> 

#> The following genes are present in one of the compared matrices, but not in the other.

#> GeneSet = ENSMUSG00000006724, ENSMUSG00000026934, ENSMUSG00000115958

#> 

#> As a result an semi join between both matrices over the gene names will be performed.

#> - adj_mat_qry: nrow = (46) and ncol(46)

#> - adj_mat_sbj: nrow = (46) and ncol(46)

#> The Hamming Distances for each gene between the two input matrices are computed.

#> 

#> Comparison 3: 'adj_mat_raw_no_noise_removed_and_no_qnorm' vs 'adj_mat_noise_removed_and_qnorm'

#> 

#> The following genes are present in one of the compared matrices, but not in the other.

#> GeneSet = ENSMUSG00000006724, ENSMUSG00000026934, ENSMUSG00000115958

#> 

#> As a result an semi join between both matrices over the gene names will be performed.

#> - adj_mat_qry: nrow = (46) and ncol(46)

#> - adj_mat_sbj: nrow = (46) and ncol(46)

#> The Hamming Distances for each gene between the two input matrices are computed.

#> 

#> Comparison 4: 'adj_mat_no_noise_removed_but_qnorm' vs 'adj_mat_noise_removed_and_qnorm'

#> 

#> The following genes are present in one of the compared matrices, but not in the other.

#> GeneSet = ENSMUSG00000006724, ENSMUSG00000026934, ENSMUSG00000115958

#> 

#> As a result an semi join between both matrices over the gene names will be performed.

#> - adj_mat_qry: nrow = (46) and ncol(46)

#> - adj_mat_sbj: nrow = (46) and ncol(46)

#> The Hamming Distances for each gene between the two input matrices are computed.

# look at results
benchmark_hamming

#> # A tibble: 46 x 5
#>    grn_tool genes     no_noise_no_qnorm_… no_noise_no_qnorm_… no_noise_with_qno…
#>    <chr>    <chr>                   <dbl>               <dbl>              <dbl>
#>  1 GENIE3   ENSMUSG0…                   2                   7                  3
#>  2 GENIE3   ENSMUSG0…                  10                  14                  5
#>  3 GENIE3   ENSMUSG0…                   1                   9                  5
#>  4 GENIE3   ENSMUSG0…                   3                   6                  4
#>  5 GENIE3   ENSMUSG0…                   2                   7                  3
#>  6 GENIE3   ENSMUSG0…                   2                   6                  3
#>  7 GENIE3   ENSMUSG0…                   7                  14                  6
#>  8 GENIE3   ENSMUSG0…                   1                   9                  7
#>  9 GENIE3   ENSMUSG0…                   4                   4                  3
#> 10 GENIE3   ENSMUSG0…                   4                   1                  0
#> # … with 36 more rows

# Run Benchmark using Jaccard Coefficients
benchmark_jaccard <- network_benchmark_noise_filtering(
genie3_49_raw,
genie3_49_noNoiseCM_raw,
genie3_49_qnorm_no_noise_removed,
genie3_49_noNoiseCM_qnorm,
dist_type = "jaccard",
grn_tool = "GENIE3")

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [11.27] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [9.95] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [10.75] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> Warning: The matrix provided as input for network_rescale() was coerced into symmetric.

#> network_make_binary() applies the median value over all values in the input matrix an uses [11.04] as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

#> 

#> Comparison 1: 'adj_mat_raw_no_noise_removed_and_no_qnorm' vs 'adj_mat_raw_noise_removed_but_no_qnorm'

#> 

#> The following genes are present in one of the compared matrices, but not in the other.

#> GeneSet = ENSMUSG00000006724, ENSMUSG00000026934, ENSMUSG00000115958

#> 

#> As a result an semi join between both matrices over the gene names will be performed.

#> - adj_mat_qry: nrow = (46) and ncol(46)

#> - adj_mat_sbj: nrow = (46) and ncol(46)

#> The Jaccard Similarity Coefficients for each gene between the two input matrices are computed.

#> 

#> Comparison 3: 'adj_mat_raw_no_noise_removed_and_no_qnorm' vs 'adj_mat_noise_removed_and_qnorm'

#> 

#> The following genes are present in one of the compared matrices, but not in the other.

#> GeneSet = ENSMUSG00000006724, ENSMUSG00000026934, ENSMUSG00000115958

#> 

#> As a result an semi join between both matrices over the gene names will be performed.

#> - adj_mat_qry: nrow = (46) and ncol(46)

#> - adj_mat_sbj: nrow = (46) and ncol(46)

#> The Jaccard Similarity Coefficients for each gene between the two input matrices are computed.

#> 

#> Comparison 4: 'adj_mat_no_noise_removed_but_qnorm' vs 'adj_mat_noise_removed_and_qnorm'

#> 

#> The following genes are present in one of the compared matrices, but not in the other.

#> GeneSet = ENSMUSG00000006724, ENSMUSG00000026934, ENSMUSG00000115958

#> 

#> As a result an semi join between both matrices over the gene names will be performed.

#> - adj_mat_qry: nrow = (46) and ncol(46)

#> - adj_mat_sbj: nrow = (46) and ncol(46)

#> The Jaccard Similarity Coefficients for each gene between the two input matrices are computed.

# look at results
benchmark_jaccard

#> # A tibble: 46 x 5
#>    grn_tool genes     no_noise_no_qnorm_… no_noise_no_qnorm_… no_noise_with_qno…
#>    <chr>    <chr>                   <dbl>               <dbl>              <dbl>
#>  1 GENIE3   ENSMUSG0…               0.92                0.741              0.875
#>  2 GENIE3   ENSMUSG0…               0.643               0.5                0.773
#>  3 GENIE3   ENSMUSG0…               0.955               0.625              0.783
#>  4 GENIE3   ENSMUSG0…               0.842               0.714              0.8  
#>  5 GENIE3   ENSMUSG0…               0.905               0.72               0.870
#>  6 GENIE3   ENSMUSG0…               0.929               0.793              0.889
#>  7 GENIE3   ENSMUSG0…               0.75                0.481              0.7  
#>  8 GENIE3   ENSMUSG0…               0.96                0.667              0.708
#>  9 GENIE3   ENSMUSG0…               0.846               0.852              0.885
#> 10 GENIE3   ENSMUSG0…               0.818               0.947              1    
#> # … with 36 more rows