Function to plot a dendrogram based on gene interaction

This function takes an adjacency matrix representing a gene regulatory network and builds a dendrogram based on the stength of gene interactions

network_plot_dendrogram(adj_mat, dist_method = "euclidean")

Arguments

adj_mat
an adjacency matrix to plot the dendrogram from.

dist_method
method to be used to calculate distances Default is euclidean.

Author

Sergio Vasquez and Hajk-Georg Drost

Examples

# path to PIDC output file
pidc_output <- system.file('beeline_examples/PIDC/outFile.txt', package = 'scNetworkR')
# parsing the output to an adjacency matrix
pidc_parsed <- pidc(pidc_output)

#> New names:
#> * `` -> ...1
#> * `` -> ...2
#> * `` -> ...3
#> * `` -> ...4
#> * `` -> ...5
#> * ...

# rescaling the matrix
rescaled <- network_rescale(pidc_parsed)

#> Error: Please provide a symmetric matrix as 'adj_mat' input for network_rescale().

# Visualize result
network_plot_dendrogram(rescaled)

#> Error in t(adj_mat[2:nrow(adj_mat), 2:ncol(adj_mat)]): object 'rescaled' not found