Calculate the naive degree distribution from binary adjacency matrices

This function calculates the degree distribution from a binary adjacency matrix.

network_statistics_degree_distribution_naive(adj_mat)

Arguments

adj_mat
a symmetrical binary adjacency matrix.

Author

Sergio Vasquez and Hajk-Georg Drost

Examples

# path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
# import PPCOR output into adjacency matrix
ppcor_parsed <- ppcor(ppcor_output)
# rescaling matrix
ppcor_rescaled <- network_rescale(ppcor_parsed)

#> It seems like your input matrix contains values of correlation coefficients range(-1,1). Please be aware that for negative values the absolute value will be taken before rescaling. We transformed all negative values to their absolute values.

# make weighted adjacency matrix binary
ppcor_binary_adj_mat <- network_make_binary(ppcor_rescaled, threshold = 70)

#> network_make_binary() applies 70 as cut-off threshold to transform the input weighted adjacency matrix into a binary adjacency matrix.

# calculate network statistics
ppcor_statistics <- network_statistics_degree_distribution_naive(ppcor_binary_adj_mat)
# look at results
ppcor_statistics

#> # A tibble: 19 x 2
#>    gene_name node_degree
#>    <chr>           <int>
#>  1 FGF9                0
#>  2 DHH                 0
#>  3 PGD2                0
#>  4 WT1mKTS             0
#>  5 CBX2                0
#>  6 AMH                 0
#>  7 UGR                 0
#>  8 SOX9                0
#>  9 GATA4               0
#> 10 NR0B1               1
#> 11 FOXL2               1
#> 12 DMRT1               2
#> 13 CTNNB1              2
#> 14 DKK1                2
#> 15 NR5A1               2
#> 16 RSPO1               3
#> 17 SRY                 3
#> 18 WNT4                3
#> 19 WT1pKTS             3